Fix Alchemist Grenadier and PFS issues
PFS replaces the Brew Potion feat with Extra Bombs. The Brew Potion
class feature should not be removed, it's just modified. Also, the
Expert Trainer class feature should not actually be replaced wither. It
was worded poorly. The class feature is still there and available for
prerequisites of other archetypes that are otherwise legal. So what is
being replaced is the ability that goes along with the class feature.
Not sure how to do that, but the TYPE.CavalierExpertTrainer clearly
should not be used.