Implemented Much of Eldritch Scion archetype
Added the proper spells known to the archetype. Spells are not
restricted; they will still act as a spellbook, but the proper spells
known will be displayed to the user.

Added selection of a bloodrager bloodline to the eldritch scion, with
bloodline Feats disabled. Bloodline powers seem to be appearing
correctly. As with proper spell regulation above, bloodline bonus spells
are NOT implemented.

Modified the casting stat from INT to CHA. and manually adjusted the
variables needed to properly calculate the Arcane Accuracy, Hasted
Assault, and Spell Shield Arcanas.