PF cleanup Cavalier's Mount should not be barred from Archetypes that replace Share Spells
This is just clean up of unused data that tries to bar animal archetypes
due to the cavalier's mount not having the share spells ability.
Particularly the Charger archetype. After review of the actual
Pathfinder Archetype rules, an archetype should be blocked if the same
ability in the base class has been replaced by another archetype. Since
Cavalier class removes the share spells ability, and an archetype has
not replaced or modified it, then the charger (and consequently other
archetypes as well that normally would replace share spells) will be
allowed. However, two archetypes that replace the otherwise non-existent
share spells ability would  not be able to be taken at the same time.
... This has not officially been errata'd or FAQ'd by Paizo. to clarify
the meaning, but my the description of Charger, it is clear that it was
always intended to be available to cavaliers.