Simplify quite a bit of code in NameGenPanel
Summary:
This simplifies the class NameGenPanel quite a bit and makes it easier
to process.
It also introduces a temporary class called Predicates which is intended
to be used until the switch to Java 1.8. Once that occurs this code can
be simplified even more by converting a couple of one-function-classes
to lambdas.

This also weakens many types to their higher level form (and then uses
<> to avoid redundant information)

This change also avoids the use Vector and replaces it with
List/ArrayList where possible.

(I know there should be a bug ID here for code cleanup but not sure what
the current one is)

Test Plan:
Opened GMGen NameGen and tried some functions
Built pcgen