DATA-3230 [PFS] Fix implementation of People of the North
This makes sure that the PFS specific mods only load if the Roleplaying
Guild Guide is loaded. It also makes sure the parent set is loaded in
the Pathdiner Society Complete multi-set.

To test this set with the minimal set of prerequisite sources you will
need PFS Core Assuption, APG, B1, UM and ISWG.