Move gradle output into a dedicated directory.
Gradle and windows do not seem to play nice with each other in ways I
don't fully understand.

This moves pcgen into a dedicated output directory instead as this is
better practice in the first place. We may need some slight changes to
other release scripts but that will happen in due course