Add configuration files for codacity (#2923)
Its kind of ridiculous that codacity doesn't have a way to point it to
the real files without including them in the project root. Its
questionable if we need codacity in the first place, but given we have
it, lets at least make it more useful than it otherwise would be.