Fix Faerie Dragon race & default kit
Fix Faerie Dragon default kit to exactly match the bestiary 3 stat block
entry. This is still not perfect, but it is much better. Now feats,
skills and spells match.

This is tested by creating a new faerie dragon from [create monster]
button and comparing the statblock form pcgen to the stat block from the
PRD.

There is still a problem that CMD doesn't show 18 vs trip due to having
4 legs. Data was missing LEGS:4 but this didn't resolve it.