Convert ParseResult.Fail to not take in the context
This means ParseResult.Fail will not capture the URI where the failure
occurred - rather the LoadContext (which gets back the ParseResult) will
have to pass the URI into the ParseResult reporting mechanisms.  This
corrects for many places where a context was not being passed into the
ParseResult.Fail at construction, so we may have been losing locations
where errors were occurring.

Note that to improve capture of location, this also increases scope of
ParseResult in some places where it was not appropriately utilized.
(BENEFIT SPELLLEVEL SPELLKNOWN and NATURALATTACKS specifically)