Interface Tokens
Leverages the Proxy system from -base

This system allows a load token to use an interface rather than the
object class (or ancestor class).  This allows multiple different,
otherwise unrelated objects to use the same load token.

This sets up for allowing Dynamic (which is not - and shouldn't be - a
CDOMObject) to handle variables.

This sets up a simplification of GlobalModifiers as well.

Overall, this also sets up the ability to do a simplification of the
token system - allowing them to ignore the context in many cases and act
directly on an interface (thus making the code clearer)