Proxy library items
This library item is designed to produce a proxy to be used for
transactions.  This will allow us to do a few changes in the main PCGen
repository:

(1) We will be able to enable tokens that operate on interfaces.  This
will allow us to have one shared token for multiple classes.  This will
be important for MODIFY/MODIFYOTHER which we want to share with Dynamic
objects, without having Dynamic objects take on the burden of being a
CDOMObject.

(2) Other tokens that duplicate today could be consolidated.

(3) The complex transaction system within the ObjectContext could be
eliminated, as the StagingProxy system here performs the same function
without the sprawl of boilerplate code.