Local Functions
Note that this cannot use lambda functions since they do not have any
(proper) support for equality in Java.

This implements the following functions:
this() is usable in MODIFY: IF AND ONLY IF it's used on an object with a
resolvable scope.  (It can't be used in the Global Modifiers file, for
example.  Currently it also can't be used on Equipment Head/Modifiers)

source() is usable in MODIFYOTHER: IF AND ONLY IF it's used on an object
with a resolvable scope.  (It can't be used in the Global Modifiers
file, for example.  Currently it also can't be used on Equipment
Head/Modifiers).  source() represents the object which contained the
MODIFYOTHER: token.

target() is usable in MODIFYOTHER: to represent the target where the
Modification is eventually applied.  (it represents the object
identified by the first two arguments to MODIFYOTHER)