Code changes for: Eliminates the built-in <none selected> Race
This moves a built-in object out of the code.

The <none selected> race is now a problem if the data team wishes to
have a required FACT/FACTSET.  Therefore, this pulls out the default
race, requiring the data to provide one.

The previous code awareness of <none selected> (including assigning it
as the Race to the PC at character creation and sorting it to the top in
the UI) has been replaced by code awareness of a GROUP called
"Unselected"

This is an incompatible change that will require the data team to
implement a Race with the appropriate characteristics (name and GROUP)