PFS season 8 updates (#4106)
* Add S07-00

* Document recovered treatures PA not working

* Hide bonus langauge tracking template from OS

* Some fixes to Recovered Treasures

* Start season 8 support

* Correct formatting of Recovered Treasures

* Restrict some boons to specific factions

* Add SOURCELONG to PFS equipment

* Remove source from PFS Chronicle list

The source company is not neccessary since all chronicle sheets are from
Paizo. The source document is not necessary since the chronicle sheet
name is the soruce document. Removing these tightens up the chronicle
sheet list.

* Most of season 8

Still missing 10, 13-17, 22, 24, 25, 99B and 99C.

* Fix season 7

Fixed season 7 to work with some changes that were implemented in season
8.

* Revert "Fix season 7"

This reverts commit 8d198d77c214795ea303067bf99fd55c34723178.