[sponsor] Clean up after all sponsors have died
As far as I could tell we only have ever had two sponsors: ourselves,
and Silven Publishing. The latter has gone the way of the dodo. The
former does not just justify adding  complexity to the lst loader, gui,
settings, and other things.

If we do get new sponsor's it'd likely make more sense to build it in a
slightly less intrusive way.