Removed mindless trait from poppet familiar
https://www.d20pfsrd.com/feats/general-feats/poppet-familiar/
https://www.d20pfsrd.com/basics-ability-scores/glossary/

By becoming a familiar the poppet gains similar aspects to an awakened construct as it gaining it an INT score and the ability to speak. 

The description for Mindless is "A creature with no Intelligence score. " At lvl 1 a PC's familiar gains an INT of 6 so it would no longer qualify for the trait.