Update installers/win-installer/pcgen.nsi
This is an attempt to correct the oowing issues:

Windows 8 and later do not suport multiple level start menu folders. i.e. PCGen\ver1\pcgen.lnk and PCGen\ver2\pcgen.lnk will be shown in the Windows 8/10 start menu as two copies of PCGen > pcgen.lnk. The user will not know which is which since the name is the same. So the top menu name needs to include the version number such as PCGen-ver1\pcgen.lnk and PCGen-ver2\pcgen.lnk. Windows will then place the two links in separate folders in the start menu.

The uninstall registry entry DisplayName property needs to be labeled exactly the same as the PCGen program link name so that the control add/remove program items entry shows the correct icon and actual installation size. And since PCGen allows multiple versions to be installed, both the uninstall registry entry DisplayName and the the PCGen link name needs to include the version number. Therefore, PCGen should use "PCGen ver1" for both the DisplayName and the start menu program link name.