ItemFacet: return an Optional from `get`
This makes it clear whether or not we're handling `null` properly. There
are several cases here that look like bugs. I've attempted
to change nothing about the semantics. If you find bugs of the sort "we
should add a check there", but there wasn't one before, please comment,
but I'll fix them in a followup.

In a couple of cases we have some odd APIs (e.g., `.get().get()`) but
IMHO the tradeoffs are worth it.