Settings: remove 'rog' setting
This setting was initially added CVS in 1.175 of SettingsHandler.java.
At present there is no UI to modifying the setting, though in theory
it is honored if it was set in the past.

It doesn't seem useful to keep it around, is undocumented, and adds a
mild amount of complexity to some tokens.