[CoreUtility] speed up file reading.
The previous code makes an attempt to be fast by reading into a
StringBuilder and using a buffer, but in reality the JVM can do a better
job than we can.

We can do even better if we were to hide the 'reader' abstraction and
handle files differently from net URLs.

- while here: fix a security issue where we default to assuming URLs are
non-net. This would break with e.g. h2 or QUIC schemes.