Allow BaseSize Code control to be used independently of PCSIZE Control
The previous design seems to have assumed they would be used together,
but that shouldn't be necessary.  Because SizeFacet's PC Size
calculation depends on the base size, we need to intercept that within
the Facet and not only do the code control interception in the case when
the base size is requested from PlayerCharacter.