[gui3] use FXML instead of creating nodes in code.
This continues to more fully decouple code from UI. Further it makes the
notion of a controller "real" in pcgen.

Long term we may even have people who edit the UI that don't need code
permissions or info.

While this is a trivial example, I'm trying to figure out the right
pattern for UI in pcgen.