SplashScreen: Convert splash screen to JavaFX
This is starting to set a pattern w.r.t. how UI is going to work. This
pattern seems like it will work well so long as we have mixed JavaFX and
Swing code. Once we only have JavaFX we'll need to rejigger parts of the
high level structure but the main code should Just Work