TabsPanel: remove it
We don't currently display this panel at all. In addition, most of what
it controls will be useless once the JavaFX conversion takes place. The
only UI customization we'll provide is what CSS allows.

This leaves the 'name display style' preference for now. It is
reasonable to make this an option, though ideally it'd be controlled by
data. I may remove the setting in a future commit, but making separate
to make distinction clear.