Tab Display Style: remove it (#5587)
As it happens, the tab display style heavily depends on the code
understanding details like "race" or "class" and can be fully replaced
by having a "tab label" in the file (which there is). If we ever want to
bring it back, we'll need a data driven list of options and format
strings.