Make available (free) memory calculation in linux use correct value for /proc/meminfo
In /proc/meminfo, the MemFree value is the amount of non allocated memory. Linux uses available memory as cache to speed-up IO, keeping only a small amount of unused memory as free memory. So the correct calculation to determine the amount of free memory should instead use the MemAvailable field.
Fixes PCGen#5796