[bugfix] handle opening debug dialog from status bar
Problem

The status bar is javafx but currently tries to open the debug dialog
through a pattern that expects the swing thread. This prevents opening
the dialog via the button (though F10 works as it is on the swing
thread).

Solution

Rejigger the mechanism of opening the debug dialog to be callable at
multiple sites. This also makes the dialog a singleton as there is no
reason for multiple instances to exist.