Adding all possible CMD choices to Human/Fighter
The text of the favored class ability, according to Archives of Nethys: https://www.aonprd.com/ClassDisplay.aspx?ItemName=Fighter

"Human ([Advanced Race Guide pg. 73], [Advanced Player's Guide pg. 23]): Add +1 to the fighter’s CMD when resisting two combat maneuvers of the character’s choice."

I've verified that the exact text is in the APG, which is the same book that introduces the new combat maneuvers.