Fixed issues with Order of the Blossom and Order of the Guard Cavaliers (#6772)
* Update amh_equip_arms_armor.lst

* Update oa_abilities_class.lst

Updated Kineticist's "Kinetic Blast" damage to scale properly, at odd levels instead of even.  Ref: Jira issue DATA-4287.

* Fixed issues with Order of the Blossom and Order of the Guard Cavaliers

Order of the Blossom's Challenge ability .MOD was not limited to just Order of the Blossom, so it was getting inserted into other cavaliers' Challenge ability descriptions.  Added a PREABILITY tag.

Order of the Guard's order description was a copy/paste of Order of the Beast.