Changes to work with Monocle 17.0.10
Commented out a bunch of constraints that are not needed.
Not sure about the runtime code that explicitly checks for
`1.8.0*` in regards to Monocle, though.  That can _probably_
go away as well, but until I spend some time looking into the
details, I can't be sure.

Added two `--add-exports` statements and a single `--add-opens`
statement.  These are nevcessary to get a couple of the tests
built properly in combination with Monocle.

Removed a bunch of targetPlatform() calls that aren't needed
for testing -- if PCGen builds on my laptop, it should build
everywhere.  (Well, almost everywhere.)  This means I don't
have to specify such things every time I do a build from the
command line.