DATA-4375 Skald's Inspired Rage Temp Bonus is missing
The Skald’s Inspired Rage is the class’s equivalent to the Bard’s Inspire Courage. As such, it would be helpful to have it appear as a temporary bonus for the players in a party with a Skald.
This Temp Bonus allows users to select their ally's Skald level so that the bonus calculates automatically.